requirements.txt
opencv-python
face_recognition
pyttsx3
scikit-learn
joblib
picamera2
RPi.GPIO
SpeechRecognition